ARM: ts7670d: Add support for TS-7670 rev D.
This is a fork of the TS-7400v2 port, largely so we can default to a
device tree that handles eMMC instead of NAND flash.

There's apparently some CPU strappings on the board that can be used to
detect which is in use, but for now, this at least gets us out of a hole
and will let us boot.  We don't think we'll be supporting the old rev B
boards for too much longer.